Unify progress bar of check and prune commands
